fix: won't display options in help if this is not declared in the command

+++ src/utils.ts
@@ -5,7 +5,7 @@ import chalk from "chalk";
 import { ok } from "./console";
 import { targetPath } from "./paths";
 import { join } from "path";
-import { mkdirSync, copyFileSync } from "fs";
+import { mkdirSync, copyFileSync, existsSync } from "fs";
 function print (mode:string, text:string) {
     let fpath = text.replace(targetPath(), '').replace(/\\/gm, '/').split('/');
     ok(`\t${chalk.cyan(mode)}: ${printRelativePath(...fpath)}`);
@@ -58,14 +58,19 @@ export function iter<T=any>(obj: any) {
 	 };
 }
 export function makeDir(location:string) {
-	mkdirSync(location);
-	ok(printRelativePath(location));
+	if(!existsSync(location)) {
+		mkdirSync(location);
+		ok(printRelativePath(location));
+	}
 }
 export function cp(source:string, target:string) {
-	copyFileSync(source, target);
-	ok(printRelativePath(target));
+	if(existsSync(source)) {
+		copyFileSync(source, target);
+		ok(printRelativePath(target));
+	}
 }
 export { override } from './override';
 export { remove } from './remove';
 export { render } from './render';
 export { spawn, spinSpawn, rawSpawn } from './spawn';
+export * from './paths';
